refactor: move entry points to testbench module
This is part of the effort to remove any references to "emulator" in the
repository. I left `emulator.py` around for backwards compatibility, but
I suspect we can remove it after checking with a few folks.
