+++ tests/test_testbench_object_metadata.py
@@ -14,22 +14,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Unit test for Object Metadata operations in emulator.py."""
+"""Unit test for Object Metadata operations in the testbench."""
 
 import json
 import os
 import unittest
 from werkzeug.test import create_environ
 
-import emulator
 import gcs
-import testbench
+from testbench import rest_server
 
 
-class TestEmulatorObjectMetadata(unittest.TestCase):
+class TestTestbenchObjectMetadata(unittest.TestCase):
     def setUp(self):
-        emulator.db.clear()
-        self.client = emulator.server.test_client()
+        rest_server.db.clear()
+        self.client = rest_server.server.test_client()
         # Avoid magic buckets in the test
         os.environ.pop("GOOGLE_CLOUD_CPP_STORAGE_TEST_BUCKET_NAME", None)
