+++ tests/test_testbench_object_special.py
@@ -14,21 +14,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Unit test for special object operations in emulator.py."""
+"""Unit test for special object operations in the testbench."""
 
 import json
 import os
 import unittest
 from werkzeug.test import create_environ
 
-import emulator
-import testbench
+from testbench import rest_server
 
 
-class TestEmulatorObjectSpecial(unittest.TestCase):
+class TestTestbenchObjectSpecial(unittest.TestCase):
     def setUp(self):
-        emulator.db.clear()
-        self.client = emulator.server.test_client()
+        rest_server.db.clear()
+        self.client = rest_server.server.test_client()
         # Avoid magic buckets in the test
         os.environ.pop("GOOGLE_CLOUD_CPP_STORAGE_TEST_BUCKET_NAME", None)
